@@ -4,7 +4,7 @@
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>com.logicommerce</groupId>
 	<artifactId>sdktest</artifactId>
-	<version>1.3.13</version>
+	<version>2.0.0</version>
 	<packaging>jar</packaging>
 	<name>${project.groupId}:${project.artifactId}</name>
 	<description>Logicommerce SDK Testing tools for Logicommerce plugins.</description>
@@ -40,7 +40,7 @@
 		</repository>
 	</distributionManagement>
 	<properties>
-		<java.version>11</java.version>
+		<java.version>17</java.version>
 		<maven.compiler.source>${java.version}</maven.compiler.source>
 		<maven.compiler.target>${java.version}</maven.compiler.target>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -51,12 +51,12 @@
 		<maven-javadoc-plugin.version>3.3.0</maven-javadoc-plugin.version>
 		<nexus-staging-maven-plugin.version>1.6.14</nexus-staging-maven-plugin.version>
 
-		<logicommerce-sdk.version>1.3.13</logicommerce-sdk.version>
-		<logicommerce-utilities.version>1.0.3</logicommerce-utilities.version>
+		<logicommerce-sdk.version>2.0.0</logicommerce-sdk.version>
+		<logicommerce-utilities.version>1.1.0</logicommerce-utilities.version>
 
-		<apache-httpclient5-version>5.1.3</apache-httpclient5-version>
+		<apache-httpclient5-version>5.3.1</apache-httpclient5-version>
 		<asynchttpclient.version>2.12.3</asynchttpclient.version>
-		<jackson.version>2.14.2</jackson.version>
+		<jackson.version>2.17.2</jackson.version>
 
 		<junit-platform-surefire-provider.version>1.3.2</junit-platform-surefire-provider.version>
 		<junit-jupiter-api.version>5.9.2</junit-jupiter-api.version>
@@ -138,20 +138,6 @@
 			</plugin>
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-gpg-plugin</artifactId>
-				<version>${maven-gpg-plugin.version}</version>
-				<executions>
-					<execution>
-						<id>sign-artifacts</id>
-						<phase>verify</phase>
-						<goals>
-							<goal>sign</goal>
-						</goals>
-					</execution>
-				</executions>
-			</plugin>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-source-plugin</artifactId>
 				<version>${maven-source-plugin.version}</version>
 				<executions>
@@ -179,17 +165,40 @@
 					</execution>
 				</executions>
 			</plugin>
-			<plugin>
-				<groupId>org.sonatype.plugins</groupId>
-				<artifactId>nexus-staging-maven-plugin</artifactId>
-				<version>${nexus-staging-maven-plugin.version}</version>
-				<extensions>true</extensions>
-				<configuration>
-					<serverId>ossrh</serverId>
-					<nexusUrl>https://s01.oss.sonatype.org/</nexusUrl>
-					<autoReleaseAfterClose>true</autoReleaseAfterClose>
-				</configuration>
-			</plugin>
 		</plugins>
 	</build>
+	<profiles>
+		<profile>
+			<id>deploy</id>
+			<build>
+				<plugins>
+					<plugin>
+						<groupId>org.apache.maven.plugins</groupId>
+						<artifactId>maven-gpg-plugin</artifactId>
+						<version>${maven-gpg-plugin.version}</version>
+						<executions>
+							<execution>
+								<id>sign-artifacts</id>
+								<phase>verify</phase>
+								<goals>
+									<goal>sign</goal>
+								</goals>
+							</execution>
+						</executions>
+					</plugin>
+					<plugin>
+						<groupId>org.sonatype.plugins</groupId>
+						<artifactId>nexus-staging-maven-plugin</artifactId>
+						<version>${nexus-staging-maven-plugin.version}</version>
+						<extensions>true</extensions>
+						<configuration>
+							<serverId>ossrh</serverId>
+							<nexusUrl>https://s01.oss.sonatype.org/</nexusUrl>
+							<autoReleaseAfterClose>true</autoReleaseAfterClose>
+						</configuration>
+					</plugin>
+				</plugins>
+			</build>
+		</profile>
+	</profiles>
 </project>
