@@ -6,13 +6,13 @@ import java.nio.ByteBuffer;
 import java.nio.charset.Charset;
 import java.util.ArrayList;
 import java.util.List;
-import io.netty.handler.codec.http.HttpHeaders;
-import io.netty.handler.codec.http.cookie.Cookie;
 import org.asynchttpclient.Response;
 import org.asynchttpclient.uri.Uri;
+import io.netty.handler.codec.http.HttpHeaders;
+import io.netty.handler.codec.http.cookie.Cookie;
 
 public class AsyncErrorResponse implements Response {
-	
+
 	private int statusCode;
 	private String responseBody;
 
@@ -118,4 +118,5 @@ public class AsyncErrorResponse implements Response {
 	public SocketAddress getLocalAddress() {
 		return null;
 	}
+
 }
